Review actions button in the toolbar

+++ applications/mail/src/app/components/toolbar/DeleteButton.tsx
@@ -12,7 +12,7 @@ import {
     Alert
 } from 'react-components';
 import { MAILBOX_LABEL_IDS } from 'proton-shared/lib/constants';
-import { deleteMessages, emptyLabel } from 'proton-shared/lib/api/messages';
+import { deleteMessages } from 'proton-shared/lib/api/messages';
 import { deleteConversations } from 'proton-shared/lib/api/conversations';
 import { c, msgid } from 'ttag';
 
@@ -23,7 +23,7 @@ import { MailSettings } from 'proton-shared/lib/interfaces';
 import { Breakpoints } from '../../models/utils';
 import { labelIncludes } from '../../helpers/labels';
 
-const { TRASH, SPAM, DRAFTS, ALL_DRAFTS, ALL_MAIL, INBOX, SENT, ALL_SENT, ARCHIVE } = MAILBOX_LABEL_IDS;
+const { TRASH, SPAM, DRAFTS, ALL_DRAFTS, SENT, ALL_SENT } = MAILBOX_LABEL_IDS;
 
 interface Props {
     labelID: string;
@@ -44,8 +44,6 @@ const DeleteButton = ({ labelID = '', mailSettings, breakpoints, selectedIDs = [
     const displayDelete =
         labelIncludes(labelID, TRASH, SPAM, DRAFTS, ALL_DRAFTS, SENT, ALL_SENT) &&
         (!breakpoints.isNarrow || !labelIncludes(labelID, SENT, ALL_SENT));
-
-    const displayEmpty = !breakpoints.isNarrow && !labelIncludes(labelID, INBOX, SENT, ALL_SENT, ARCHIVE, ALL_MAIL);
 
     const handleDelete = async () => {
         await new Promise((resolve, reject) => {
@@ -72,49 +70,19 @@ const DeleteButton = ({ labelID = '', mailSettings, breakpoints, selectedIDs = [
         createNotification({ text: c('Success').t`Elements deleted` });
     };
 
-    const handleEmpty = async () => {
-        await new Promise((resolve, reject) => {
-            createModal(
-                <ConfirmModal
-                    title={c('Title').t`Empty folder`}
-                    confirm={(<ErrorButton type="submit" icon={null}>{c('Action').t`Empty`}</ErrorButton>) as any}
-                    onConfirm={resolve}
-                    onClose={reject}
-                >
-                    <Alert type="warning">{c('Info')
-                        .t`This action will permanently delete your emails. Are you sure you want to empty this folder?`}</Alert>
-                </ConfirmModal>
-            );
-        });
-        c;
-        await api(emptyLabel({ LabelID: labelID, AddressID: undefined }));
-        await call();
-        createNotification({ text: c('Success').t`Folder cleared` });
-    };
+    if (!displayDelete) {
+        return null;
+    }
 
     return (
-        <>
-            {displayDelete ? (
-                <ToolbarButton
-                    loading={loading}
-                    title={c('Action').t`Delete`}
-                    onClick={() => withLoading(handleDelete())}
-                    disabled={!selectedIDs.length}
-                >
-                    <Icon className="toolbar-icon mauto" name="delete" />
-                </ToolbarButton>
-            ) : null}
-            {displayEmpty ? (
-                <ToolbarButton
-                    loading={loading}
-                    title={c('Action').t`Empty folder`}
-                    onClick={() => withLoading(handleEmpty())}
-                    disabled={!selectedIDs.length}
-                >
-                    <Icon className="toolbar-icon mauto" name="empty-folder" />
-                </ToolbarButton>
-            ) : null}
-        </>
+        <ToolbarButton
+            loading={loading}
+            title={c('Action').t`Delete`}
+            onClick={() => withLoading(handleDelete())}
+            disabled={!selectedIDs.length}
+        >
+            <Icon className="toolbar-icon mauto" name="delete" />
+        </ToolbarButton>
     );
 };
